feat: Add admin layout with sidebar, header, and user navigation components
