+++ app/(dashboard)/admin/conges/page.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const page = () => {
+  return <div></div>;
+};
+
+export default page;
