+++ app/globals.css
@@ -80,7 +80,7 @@
   --chart-3: oklch(0.398 0.07 227.392);
   --chart-4: oklch(0.828 0.189 84.429);
   --chart-5: oklch(0.769 0.188 70.08);
-  --sidebar: oklch(0.985 0 0);
+  --sidebar: oklch(100% 0.00011 271.152);
   --sidebar-foreground: oklch(0.145 0 0);
   --sidebar-primary: oklch(0.205 0 0);
   --sidebar-primary-foreground: oklch(0.985 0 0);
@@ -93,7 +93,7 @@
 @layer base {
   * {
     @apply border-border outline-ring/50;
-    font-family: var(--font-urbanist);
+    font-family: var(--font-manrope);
     /* border: 1px solid red; */
   }
   body {
